+++ apps/admin_panel/assets/package.json
@@ -48,6 +48,7 @@
     "fetch-mock": "^6.0.0",
     "file-loader": "^1.1.5",
     "html-webpack-plugin": "^2.30.1",
+    "identity-obj-proxy": "^3.0.0",
     "jest": "^22.0.1",
     "node-sass": "^4.7.2",
     "react-test-renderer": "^16.2.0",
