+++ apps/admin_panel/assets/src/omisego/__tests__/stubs/params.stubs.js
@@ -3,5 +3,5 @@ import { OMISEGO_BASE_URL } from '../../config';
 export const loginValidParams = { email: 'email@example.com', password: 'password' };
 export const loginInvalidParams = { email: 'email', password: 'pass' };
 
-export const forgotPasswordValidParams = { email: 'email@example.com', url: OMISEGO_BASE_URL };
-export const forgotPasswordInvalidParams = { email: 'email', url: '' };
+export const resetPasswordValidParams = { email: 'email@example.com', url: OMISEGO_BASE_URL };
+export const resetPasswordInvalidParams = { email: 'email', url: '' };
