+++ apps/admin_panel/assets/src/omisego/config.js
@@ -1,9 +1,9 @@
 /* Euro's config */
-export const OMISEGO_BASE_URL = 'http://10.5.10.20:5000/';
-export const OMISEGO_API_KEY_ID = '566c55d8-b0dc-4663-bcfd-30b50cc25530';
-export const OMISEGO_API_KEY = 'LF2M1hyOaO4S_MUGlsoeECV3O6qAyLIFEG0Du2yePFk';
+// export const OMISEGO_BASE_URL = 'http://10.5.10.20:5000/';
+// export const OMISEGO_API_KEY_ID = '4162ab86-4e62-49d6-a6fb-aa39f0cf4542';
+// export const OMISEGO_API_KEY = 'YkvHuKCHE6qx89bY_z3aUaxRt8OU9xpn92pxnwg6ipk';
 
 /* Mederic's config */
-// export const OMISEGO_BASE_URL = 'http://localhost:5000/';
-// export const OMISEGO_API_KEY_ID = 'b64ad9c6-81b6-436d-8b17-b0a97a8b62c0';
-// export const OMISEGO_API_KEY = 'mzLD8PqbL3uWSB8I8Pp6servdxMz45LR_S9sKixx3R8';
+export const OMISEGO_BASE_URL = 'http://localhost:5000/';
+export const OMISEGO_API_KEY_ID = '924d5fec-5ecc-4083-8414-0f14529d5c7a';
+export const OMISEGO_API_KEY = 'UMGvF3Q9UAx0WZQ3oblXsPAwx4U8BLEk8Fje67BxGQ8';
