--- apps/admin_panel/assets/src/omisego/services/create_new_admin_api.js
@@ -1,19 +1,0 @@
-import request from './api_service';
-
-export default function createAdmin(params, callback) {
-  const {
-    email, token, password, passwordConfirm,
-  } = params;
-  const requestParams = {
-    path: 'invite.accept',
-    params: {
-      email,
-      token,
-      password,
-      password_confirm: passwordConfirm,
-    },
-    authenticated: false,
-    callback,
-  };
-  return request(requestParams);
-}
