+++ apps/admin_panel/assets/src/omisego/services/session_api.js
@@ -20,25 +20,30 @@ export function logout(callback) {
   return request(requestParams);
 }
 
-export function forgotPassword(params, callback) {
-  callback(null, {});
-  // const requestParams = {
-  //   path: 'forgot_password',
-  //   params: JSON.stringify(params),
-  //   authenticated: false,
-  //   callback,
-  // };
-  // return request(requestParams);
+export function resetPassword(params, callback) {
+  const {
+    email, url,
+  } = params;
+  const requestParams = {
+    path: 'password.reset',
+    params: JSON.stringify({ email, redirect_url: url }),
+    authenticated: false,
+    callback,
+  };
+  return request(requestParams);
 }
 
-export function resetPassword(params, callback) {
-  callback(null, {});
-  // const { resetToken, ...rest } = params;
-  // return request(
-  //   'reset_password',
-  //   JSON.stringify({
-  //     reset_token: resetToken, ...rest,
-  //   }),
-  //   callback,
-  // );
+export function updatePassword(params, callback) {
+  const {
+    resetToken, password, passwordConfirmation, email,
+  } = params;
+  const requestParams = {
+    path: 'password.update',
+    params: JSON.stringify({
+      email, token: resetToken, password, password_confirmation: passwordConfirmation,
+    }),
+    authenticated: false,
+    callback,
+  };
+  return request(requestParams);
 }
