+++ apps/admin_panel/assets/src/pages/App.jsx
@@ -5,15 +5,17 @@ import { ConnectedRouter } from 'react-router-redux';
 import PropTypes from 'prop-types';
 import AlertActions from '../actions/alert.actions';
 import history from '../helpers/history';
-import { invitationConst } from '../omisego/services/setting_api';
+
+import { UPDATE_PASSWORD } from './public/reset_password/ResetPasswordForm';
+import { INVITATION } from './authenticated/setting/Setting';
 
 import DevTools from './DevTools';
 import AccountRouter from './authenticated/AccountRouter';
 import PublicRoute from './public/PublicRoute';
 import Startup from './startup/Startup';
 import SignIn from './public/signin/SignIn';
-import ForgotPassword from './public/forgot_password/ForgotPassword';
 import ResetPassword from './public/reset_password/ResetPassword';
+import UpdatePassword from './public/update_password/UpdatePassword';
 
 class App extends Component {
   constructor(props) {
@@ -34,9 +36,9 @@ class App extends Component {
           <div className="container">
             <Switch>
               <PublicRoute component={SignIn} path="/signin" />
-              <PublicRoute component={ForgotPassword} path="/forgot_password" />
               <PublicRoute component={ResetPassword} path="/reset_password" />
-              <PublicRoute component={ResetPassword} path={`/${invitationConst.pathname}`} />
+              <PublicRoute component={UpdatePassword} path={`/${UPDATE_PASSWORD.pathname}`} />
+              <PublicRoute component={UpdatePassword} path={`/${INVITATION.pathname}`} />
               <AccountRouter path="/a/:id" />
               <AccountRouter path="/" />
               <Redirect to="/signin" />
