+++ apps/admin_panel/assets/src/pages/public/reset_password/__tests__/ResetPasswordForm.snapshot.test.jsx
@@ -1,12 +1,12 @@
 import renderer from 'react-test-renderer';
-import ForgotPasswordForm from '../ForgotPasswordForm';
+import ResetPasswordForm from '../ResetPasswordForm';
 import { SnapshotComponentTester } from '../../../../../__tests__/ComponentTester';
 
 it('renders correctly', () => {
-  const forgotPassword = jest.fn();
+  const resetPassword = jest.fn();
   const onSuccess = jest.fn();
-  const tree = renderer.create(SnapshotComponentTester(ForgotPasswordForm, {
-    loading: false, forgotPassword, onSuccess,
+  const tree = renderer.create(SnapshotComponentTester(ResetPasswordForm, {
+    loading: false, resetPassword, onSuccess,
   })).toJSON();
   expect(tree).toMatchSnapshot();
 });
