+++ apps/admin_panel/assets/src/pages/public/reset_password/__tests__/ResetPasswordSuccess.test.jsx
@@ -1,18 +1,14 @@
-import Enzyme from 'enzyme';
-import Adapter from 'enzyme-adapter-react-16';
 import { createMemoryHistory } from 'history';
-import ForgotPasswordSuccess from '../ForgotPasswordSuccess';
+import ResetPasswordSuccess from '../ResetPasswordSuccess';
 import { ShallowComponentTester } from '../../../../../__tests__/ComponentTester';
 
-Enzyme.configure({ adapter: new Adapter() });
-
-describe('ForgotPasswordSuccess', () => {
+describe('ResetPasswordSuccess', () => {
   test('redirect to signin on click', () => {
     const history = createMemoryHistory();
     // Note that the shallow rendering doesn't let us access to the props initially
     // passed (in our case history and translate). To be able to test the redirect
     // we need to hold a reference to `history`.
-    const component = ShallowComponentTester(ForgotPasswordSuccess, { history });
+    const component = ShallowComponentTester(ResetPasswordSuccess, { history });
     expect(history.location.pathname).toBe('/');
     component.find('Button').at(0).simulate('click');
     expect(history.location.pathname).toBe('/signin');
