+++ apps/admin_panel/assets/src/pages/public/reset_password/__tests__/actions.test.js
@@ -15,31 +15,31 @@ jest.mock('../../../../helpers/errorHandler', () => (
   { handleAPIError: jest.fn() }
 ));
 
-describe('forgotPassword action', () => {
+describe('resetPassword action', () => {
   test('shows and hide loading when requesting', () => {
     const callback = jest.fn();
-    sessionAPI.forgotPassword = jest.fn((params, cb) => cb(null, validResponse));
+    sessionAPI.resetPassword = jest.fn((params, cb) => cb(null, validResponse));
     const store = mockStore();
     const expectedActions = [
       { type: globalConstants.SHOW_LOADING },
       { type: globalConstants.HIDE_LOADING },
     ];
-    store.dispatch(Actions.forgotPassword(validParams, callback));
+    store.dispatch(Actions.resetPassword(validParams, callback));
     expect(store.getActions()).toEqual(expect.arrayContaining(expectedActions));
   });
 
   test('call onSuccess when success', () => {
     const callback = jest.fn();
-    sessionAPI.forgotPassword = jest.fn((params, cb) => cb(null, validResponse));
+    sessionAPI.resetPassword = jest.fn((params, cb) => cb(null, validResponse));
     const store = mockStore();
-    store.dispatch(Actions.forgotPassword(validParams, callback));
+    store.dispatch(Actions.resetPassword(validParams, callback));
     expect(callback).toBeCalled();
   });
 
   test('calls handleAPIError when failing', () => {
     const callback = jest.fn();
-    sessionAPI.forgotPassword = jest.fn((params, cb) => cb(errorResponse, null));
-    mockStore().dispatch(Actions.forgotPassword(validParams, callback));
+    sessionAPI.resetPassword = jest.fn((params, cb) => cb(errorResponse, null));
+    mockStore().dispatch(Actions.resetPassword(validParams, callback));
     expect(ErrorHandler.handleAPIError).toBeCalledWith(expect.any(Function), errorResponse);
   });
 });
