+++ apps/admin_panel/assets/src/pages/public/reset_password/__tests__/stubs.js
@@ -1,4 +1,5 @@
-import { OMISEGO_BASE_URL } from '../../../../omisego/config';
+import { formatEmailLink } from '../../../../helpers/urlFormatter';
+import { INVITATION } from '../../../authenticated/setting/Setting';
 
 export const validResponse = {};
 export const errorResponse = {
@@ -9,4 +10,4 @@ export const errorResponse = {
 };
 export const validEmail = 'email@example.com';
 export const invalidEmail = 'invalidemail';
-export const validParams = { url: OMISEGO_BASE_URL, email: validEmail };
+export const validParams = { redirect_url: formatEmailLink(INVITATION), email: validEmail };
