+++ apps/admin_panel/assets/src/pages/public/update_password/UpdatePasswordSuccess.jsx
@@ -5,27 +5,25 @@ import { getTranslate } from 'react-localize-redux';
 import PropTypes from 'prop-types';
 import { withRouter } from 'react-router-dom';
 
-const ForgotPasswordSuccess = ({ translate, history }) => (
+const ResetPasswordSuccess = ({ successText, translate, history }) => (
   <div className="omg-form">
     <h2 className="omg-form__title">
-      {translate('forgot-password-success.reset_password_complete')}
+      {successText}
     </h2>
-    <h3 className="omg-form__subtitle">
-      {translate('forgot-password-success.please_check_your_email')}
-    </h3>
     <Button
       bsClass="btn btn-omg-blue"
       bsStyle="primary"
       onClick={() => { history.push('/signin'); }}
     >
-      {translate('forgot-password-success.back_to_sign_in')}
+      {translate('update-password-success.back_to_sign_in')}
     </Button>
   </div>
 );
 
 
-ForgotPasswordSuccess.propTypes = {
+ResetPasswordSuccess.propTypes = {
   history: PropTypes.object.isRequired,
+  successText: PropTypes.string.isRequired,
   translate: PropTypes.func.isRequired,
 };
 
@@ -36,4 +34,4 @@ function mapStateToProps(state) {
   };
 }
 
-export default withRouter(connect(mapStateToProps)(ForgotPasswordSuccess));
+export default withRouter(connect(mapStateToProps)(ResetPasswordSuccess));
